move game to components
