+++ packages/client/src/components/Game/Apple/index.ts
@@ -1,5 +1,6 @@
-import { GRID_SIZE, SPRITE_GRID_SIZE, GRID_COUNT } from './constants'
-import { getRandomInt } from './utils/getRandomInt'
+import { GRID_SIZE, SPRITE_GRID_SIZE, GRID_COUNT } from '../constants'
+import { getRandomInt } from '../utils/getRandomInt'
+
 export class Apple {
   x = GRID_SIZE * 20
   y = GRID_SIZE * 20
@@ -7,7 +8,7 @@ export class Apple {
   draw(context: CanvasRenderingContext2D, sprites: HTMLImageElement) {
     context.drawImage(
       sprites,
-      0 * SPRITE_GRID_SIZE,
+      0,
       3 * SPRITE_GRID_SIZE,
       SPRITE_GRID_SIZE,
       SPRITE_GRID_SIZE,
