+++ packages/client/src/components/Game/images/sand.png
