+++ packages/client/src/components/Game/images/snake-graphics.png
