+++ packages/client/src/components/Game/style.css
@@ -1,0 +1,3 @@
+.game-board {
+  border: 1px solid white;
+}
