+++ packages/client/src/pages/Game/style.css
@@ -4,7 +4,3 @@
   justify-content: center;
   box-sizing: border-box;
 }
-
-canvas {
-  border: 1px solid white;
-}
